@@ -38,15 +38,15 @@ namespace OSharp.AspNetCore.Mvc
         /// <returns></returns>
         public override IServiceCollection AddServices(IServiceCollection services)
         {
-            services.AddMvc(options =>
+            services.AddControllersWithViews(options =>
             {
                 options.Conventions.Add(new DashedRoutingConvention());
-                options.Filters.Add(new OnlineUserAuthorizationFilter()); // 构建在线用户信息
-                options.Filters.Add(new FunctionAuthorizationFilter()); // 全局功能权限过滤器
-            }).AddJsonOptions(options =>
+                options.Filters.Add(new OnlineUserAuthorizationFilter()); //构建在线用户信息
+                options.Filters.Add(new FunctionAuthorizationFilter()); //全局功能权限检查过滤器
+            }).AddNewtonsoftJson(options =>
             {
                 options.SerializerSettings.ContractResolver = new DefaultContractResolver();
-            }).SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            });
 
             services.AddScoped<UnitOfWorkFilterImpl>();
             services.AddHttpsRedirection(opts => opts.HttpsPort = 443);
@@ -61,7 +61,10 @@ namespace OSharp.AspNetCore.Mvc
         /// <param name="app">应用程序构建器</param>
         public override void UsePack(IApplicationBuilder app)
         {
-            app.UseMvcWithAreaRoute();
+            //app.UseMvcWithAreaRoute();
+            app
+                .UseRouting()
+                .UseEndpointsWithAreaRoute();
             IsEnabled = true;
         }
     }
