@@ -8,7 +8,8 @@
 // -----------------------------------------------------------------------
 
 using System.Data;
-using System.Data.SqlClient;
+
+using Microsoft.Data.SqlClient;
 
 
 namespace OSharp.Entity.SqlServer
