@@ -8,8 +8,10 @@
 // -----------------------------------------------------------------------
 
 using System;
+using System.Text.Json;
 
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -61,8 +63,8 @@ namespace OSharp.AspNetCore.SignalR
         /// <param name="app">Asp应用程序构建器</param>
         public override void UsePack(IApplicationBuilder app)
         {
-            Action<HubRouteBuilder> hubRouteBuildAction = GetHubRouteBuildAction(app.ApplicationServices);
-            app.UseSignalR(hubRouteBuildAction);
+            Action<IEndpointRouteBuilder> endpointBuilderAction = GetEndpointRouteBuilderAction(app.ApplicationServices);
+            app.UseEndpoints(endpointBuilderAction);
         }
 
         /// <summary>
@@ -72,14 +74,15 @@ namespace OSharp.AspNetCore.SignalR
         /// <returns></returns>
         protected virtual Action<ISignalRServerBuilder> GetSignalRServerBuildAction(IServiceCollection services)
         {
-            return builder => builder.AddJsonProtocol(config => config.PayloadSerializerSettings.ContractResolver = new DefaultContractResolver());
+            return builder => builder.AddNewtonsoftJsonProtocol(options =>
+                options.PayloadSerializerSettings.ContractResolver = new DefaultContractResolver());
         }
 
         /// <summary>
-        /// 重写以获取Hub路由创建委托
+        /// 重写以获取Hub终结点路由创建委托
         /// </summary>
         /// <param name="serviceProvider">服务提供者</param>
         /// <returns></returns>
-        protected abstract Action<HubRouteBuilder> GetHubRouteBuildAction(IServiceProvider serviceProvider);
+        protected abstract Action<IEndpointRouteBuilder> GetEndpointRouteBuilderAction(IServiceProvider serviceProvider);
     }
 }
