@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Security.Claims;
 using System.Text.Encodings.Web;
+using System.Text.Json;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
@@ -52,21 +53,24 @@ namespace OSharp.Identity.OAuth2.QQ
                 throw new HttpRequestException($"未能检索QQ Connect的用户信息(返回状态码:{userInformationResponse.StatusCode})，请检查参数是否正确。");
             }
 
-            JObject payload = JObject.Parse(await userInformationResponse.Content.ReadAsStringAsync());
+            string json2 = await userInformationResponse.Content.ReadAsStringAsync();
+            JObject payload = JObject.Parse(json2);
+            JsonDocument document = JsonDocument.Parse(json2);
             //identity.AddClaim(new Claim(ClaimTypes.NameIdentifier, openId, ClaimValueTypes.String, Options.ClaimsIssuer));
             payload.Add("openid", openId);
-            OAuthCreatingTicketContext context = new OAuthCreatingTicketContext(new ClaimsPrincipal(identity), properties, Context, Scheme, Options, Backchannel, tokens, payload);
+            OAuthCreatingTicketContext context = new OAuthCreatingTicketContext(new ClaimsPrincipal(identity), properties, Context, Scheme, Options, Backchannel, tokens, document.RootElement);
             context.RunClaimActions();
             await Events.CreatingTicket(context);
             return new AuthenticationTicket(context.Principal, context.Properties, Scheme.Name);
         }
 
-        /// <summary>
-        /// 通过Authorization Code获取Access Token。
-        /// 重写改方法，QQ这一步用的是Get请求，base中用的是Post
-        /// </summary>
-        protected override async Task<OAuthTokenResponse> ExchangeCodeAsync(string code, string redirectUri)
+        #region Overrides of OAuthHandler<QQOptions>
+
+        protected override async Task<OAuthTokenResponse> ExchangeCodeAsync(OAuthCodeExchangeContext context)
         {
+            string code = context.Code;
+            string redirectUri = context.RedirectUri;
+
             Dictionary<string, string> parameters = new Dictionary<string, string>
             {
                 {  "grant_type", "authorization_code" },
@@ -83,9 +87,13 @@ namespace OSharp.Identity.OAuth2.QQ
             {
                 return OAuthTokenResponse.Failed(new Exception("获取QQ Connect Access Token出错。"));
             }
-            JObject payload = JObject.Parse(企鹅的返回不拘一格传入这里统一转换为JSON(await response.Content.ReadAsStringAsync()));
-            return OAuthTokenResponse.Success(payload);
+
+            string json = 企鹅的返回不拘一格传入这里统一转换为JSON(await response.Content.ReadAsStringAsync());
+            JsonDocument document = JsonDocument.Parse(json);
+            return OAuthTokenResponse.Success(document);
         }
+
+        #endregion
 
         //Convert to JSON
         private string 企鹅的返回不拘一格传入这里统一转换为JSON(string text)
