@@ -10,31 +10,37 @@ using OSharp.Entity;
 namespace Liuliu.Demo.Web.Migrations
 {
     [DbContext(typeof(DefaultDbContext))]
-    [Migration("20190910142209_Message")]
-    partial class Message
+    [Migration("20190928051613_Init")]
+    partial class Init
     {
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.2.6-servicing-10079")
+                .HasAnnotation("ProductVersion", "3.0.0")
                 .HasAnnotation("Relational:MaxIdentifierLength", 128)
                 .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.LoginLog", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<string>("Ip");
+                    b.Property<string>("Ip")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime?>("LogoutTime");
+                    b.Property<DateTime?>("LogoutTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<string>("UserAgent");
+                    b.Property<string>("UserAgent")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -47,14 +53,18 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Name")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int?>("ParentId");
+                    b.Property<int?>("ParentId")
+                        .HasColumnType("int");
 
-                    b.Property<string>("Remark");
+                    b.Property<string>("Remark")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -67,32 +77,44 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("ConcurrencyStamp")
-                        .IsConcurrencyToken();
+                        .IsConcurrencyToken()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<DateTime?>("DeletedTime");
+                    b.Property<DateTime?>("DeletedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<bool>("IsAdmin");
+                    b.Property<bool>("IsAdmin")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsDefault");
+                    b.Property<bool>("IsDefault")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsSystem");
+                    b.Property<bool>("IsSystem")
+                        .HasColumnType("bit");
 
-                    b.Property<Guid?>("MessageId");
+                    b.Property<Guid?>("MessageId")
+                        .HasColumnType("uniqueidentifier");
 
                     b.Property<string>("Name")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
                     b.Property<string>("NormalizedName")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(450)");
 
                     b.Property<string>("Remark")
+                        .HasColumnType("nvarchar(512)")
                         .HasMaxLength(512);
 
                     b.HasKey("Id");
@@ -126,13 +148,17 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<string>("ClaimType");
+                    b.Property<string>("ClaimType")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("ClaimValue");
+                    b.Property<string>("ClaimValue")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("RoleId");
+                    b.Property<int>("RoleId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -145,54 +171,77 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int>("AccessFailedCount");
+                    b.Property<int>("AccessFailedCount")
+                        .HasColumnType("int");
 
                     b.Property<string>("ConcurrencyStamp")
-                        .IsConcurrencyToken();
+                        .IsConcurrencyToken()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<DateTime?>("DeletedTime");
+                    b.Property<DateTime?>("DeletedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<string>("Email");
+                    b.Property<string>("Email")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<bool>("EmailConfirmed");
+                    b.Property<bool>("EmailConfirmed")
+                        .HasColumnType("bit");
 
-                    b.Property<string>("HeadImg");
+                    b.Property<string>("HeadImg")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsSystem");
+                    b.Property<bool>("IsSystem")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("LockoutEnabled");
+                    b.Property<bool>("LockoutEnabled")
+                        .HasColumnType("bit");
 
-                    b.Property<DateTimeOffset?>("LockoutEnd");
+                    b.Property<DateTimeOffset?>("LockoutEnd")
+                        .HasColumnType("datetimeoffset");
 
-                    b.Property<Guid?>("MessageId");
+                    b.Property<Guid?>("MessageId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("NickName");
+                    b.Property<string>("NickName")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("NormalizeEmail");
+                    b.Property<string>("NormalizeEmail")
+                        .HasColumnType("nvarchar(450)");
 
                     b.Property<string>("NormalizedUserName")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<string>("PasswordHash");
+                    b.Property<string>("PasswordHash")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("PhoneNumber");
+                    b.Property<string>("PhoneNumber")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<bool>("PhoneNumberConfirmed");
+                    b.Property<bool>("PhoneNumberConfirmed")
+                        .HasColumnType("bit");
 
-                    b.Property<string>("Remark");
+                    b.Property<string>("Remark")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("SecurityStamp");
+                    b.Property<string>("SecurityStamp")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<bool>("TwoFactorEnabled");
+                    b.Property<bool>("TwoFactorEnabled")
+                        .HasColumnType("bit");
 
                     b.Property<string>("UserName")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -213,14 +262,18 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("ClaimType")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("ClaimValue");
+                    b.Property<string>("ClaimValue")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -233,11 +286,14 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<string>("RegisterIp");
+                    b.Property<string>("RegisterIp")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -250,19 +306,26 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserLogin", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("Avatar");
+                    b.Property<string>("Avatar")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<string>("LoginProvider");
+                    b.Property<string>("LoginProvider")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<string>("ProviderDisplayName");
+                    b.Property<string>("ProviderDisplayName")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("ProviderKey");
+                    b.Property<string>("ProviderKey")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -279,17 +342,23 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserRole", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<DateTime?>("DeletedTime");
+                    b.Property<DateTime?>("DeletedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<int>("RoleId");
+                    b.Property<int>("RoleId")
+                        .HasColumnType("int");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -306,15 +375,20 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserToken", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("LoginProvider");
+                    b.Property<string>("LoginProvider")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<string>("Name");
+                    b.Property<string>("Name")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
-                    b.Property<string>("Value");
+                    b.Property<string>("Value")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -329,33 +403,46 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Infos.Entities.Message", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<DateTime?>("BeginDate");
+                    b.Property<DateTime?>("BeginDate")
+                        .HasColumnType("datetime2");
 
-                    b.Property<bool>("CanReply");
+                    b.Property<bool>("CanReply")
+                        .HasColumnType("bit");
 
                     b.Property<string>("Content")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<DateTime?>("DeletedTime");
+                    b.Property<DateTime?>("DeletedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<DateTime?>("EndDate");
+                    b.Property<DateTime?>("EndDate")
+                        .HasColumnType("datetime2");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsSended");
+                    b.Property<bool>("IsSended")
+                        .HasColumnType("bit");
 
-                    b.Property<int>("MessageType");
+                    b.Property<int>("MessageType")
+                        .HasColumnType("int");
 
-                    b.Property<int>("NewReplyCount");
+                    b.Property<int>("NewReplyCount")
+                        .HasColumnType("int");
 
-                    b.Property<int>("SenderId");
+                    b.Property<int>("SenderId")
+                        .HasColumnType("int");
 
                     b.Property<string>("Title")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -367,17 +454,23 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Infos.Entities.MessageReceive", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<Guid>("MessageId");
+                    b.Property<Guid>("MessageId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<int>("NewReplyCount");
+                    b.Property<int>("NewReplyCount")
+                        .HasColumnType("int");
 
-                    b.Property<DateTime>("ReadDate");
+                    b.Property<DateTime>("ReadDate")
+                        .HasColumnType("datetime2");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -391,26 +484,36 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Infos.Entities.MessageReply", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<Guid>("BelongMessageId");
+                    b.Property<Guid>("BelongMessageId")
+                        .HasColumnType("uniqueidentifier");
 
                     b.Property<string>("Content")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<DateTime?>("DeletedTime");
+                    b.Property<DateTime?>("DeletedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsRead");
+                    b.Property<bool>("IsRead")
+                        .HasColumnType("bit");
 
-                    b.Property<Guid>("ParentMessageId");
+                    b.Property<Guid>("ParentMessageId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<Guid>("ParentReplyId");
+                    b.Property<Guid>("ParentReplyId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -428,19 +531,26 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.EntityRole", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<Guid>("EntityId");
+                    b.Property<Guid>("EntityId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("FilterGroupJson");
+                    b.Property<string>("FilterGroupJson")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<int>("Operation");
+                    b.Property<int>("Operation")
+                        .HasColumnType("int");
 
-                    b.Property<int>("RoleId");
+                    b.Property<int>("RoleId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -456,17 +566,23 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.EntityUser", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<Guid>("EntityId");
+                    b.Property<Guid>("EntityId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("FilterGroupJson");
+                    b.Property<string>("FilterGroupJson")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -482,21 +598,28 @@ namespace Liuliu.Demo.Web.Migrations
                 {
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Code")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
                     b.Property<string>("Name")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<double>("OrderCode");
+                    b.Property<double>("OrderCode")
+                        .HasColumnType("float");
 
-                    b.Property<int?>("ParentId");
+                    b.Property<int?>("ParentId")
+                        .HasColumnType("int");
 
-                    b.Property<string>("Remark");
+                    b.Property<string>("Remark")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("TreePathString");
+                    b.Property<string>("TreePathString")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -519,11 +642,14 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.ModuleFunction", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<Guid>("FunctionId");
+                    b.Property<Guid>("FunctionId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<int>("ModuleId");
+                    b.Property<int>("ModuleId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -539,11 +665,14 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.ModuleRole", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<int>("ModuleId");
+                    b.Property<int>("ModuleId")
+                        .HasColumnType("int");
 
-                    b.Property<int>("RoleId");
+                    b.Property<int>("RoleId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -559,13 +688,17 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.ModuleUser", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<bool>("Disabled");
+                    b.Property<bool>("Disabled")
+                        .HasColumnType("bit");
 
-                    b.Property<int>("ModuleId");
+                    b.Property<int>("ModuleId")
+                        .HasColumnType("int");
 
-                    b.Property<int>("UserId");
+                    b.Property<int>("UserId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -581,17 +714,23 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Systems.Entities.AuditEntity", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("EntityKey");
+                    b.Property<string>("EntityKey")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("Name");
+                    b.Property<string>("Name")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("OperateType");
+                    b.Property<int>("OperateType")
+                        .HasColumnType("int");
 
-                    b.Property<Guid>("OperationId");
+                    b.Property<Guid>("OperationId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("TypeName");
+                    b.Property<string>("TypeName")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -603,31 +742,44 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Systems.Entities.AuditOperation", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("Browser");
+                    b.Property<string>("Browser")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<DateTime>("CreatedTime");
+                    b.Property<DateTime>("CreatedTime")
+                        .HasColumnType("datetime2");
 
-                    b.Property<int>("Elapsed");
+                    b.Property<int>("Elapsed")
+                        .HasColumnType("int");
 
-                    b.Property<string>("FunctionName");
+                    b.Property<string>("FunctionName")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("Ip");
+                    b.Property<string>("Ip")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("Message");
+                    b.Property<string>("Message")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("NickName");
+                    b.Property<string>("NickName")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("OperationSystem");
+                    b.Property<string>("OperationSystem")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("ResultType");
+                    b.Property<int>("ResultType")
+                        .HasColumnType("int");
 
-                    b.Property<string>("UserAgent");
+                    b.Property<string>("UserAgent")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("UserId");
+                    b.Property<string>("UserId")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("UserName");
+                    b.Property<string>("UserName")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -637,19 +789,26 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("Liuliu.Demo.Systems.Entities.AuditProperty", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<Guid>("AuditEntityId");
+                    b.Property<Guid>("AuditEntityId")
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<string>("DataType");
+                    b.Property<string>("DataType")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("DisplayName");
+                    b.Property<string>("DisplayName")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("FieldName");
+                    b.Property<string>("FieldName")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("NewValue");
+                    b.Property<string>("NewValue")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("OriginalValue");
+                    b.Property<string>("OriginalValue")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -661,18 +820,23 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("OSharp.Core.EntityInfos.EntityInfo", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<bool>("AuditEnabled");
+                    b.Property<bool>("AuditEnabled")
+                        .HasColumnType("bit");
 
                     b.Property<string>("Name")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
                     b.Property<string>("PropertyJson")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
                     b.Property<string>("TypeName")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(450)");
 
                     b.HasKey("Id");
 
@@ -686,33 +850,47 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("OSharp.Core.Functions.Function", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<int>("AccessType");
+                    b.Property<int>("AccessType")
+                        .HasColumnType("int");
 
-                    b.Property<string>("Action");
+                    b.Property<string>("Action")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<string>("Area");
+                    b.Property<string>("Area")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<bool>("AuditEntityEnabled");
+                    b.Property<bool>("AuditEntityEnabled")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("AuditOperationEnabled");
+                    b.Property<bool>("AuditOperationEnabled")
+                        .HasColumnType("bit");
 
-                    b.Property<int>("CacheExpirationSeconds");
+                    b.Property<int>("CacheExpirationSeconds")
+                        .HasColumnType("int");
 
-                    b.Property<string>("Controller");
+                    b.Property<string>("Controller")
+                        .HasColumnType("nvarchar(450)");
 
-                    b.Property<bool>("IsAccessTypeChanged");
+                    b.Property<bool>("IsAccessTypeChanged")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsAjax");
+                    b.Property<bool>("IsAjax")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsCacheSliding");
+                    b.Property<bool>("IsCacheSliding")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsController");
+                    b.Property<bool>("IsController")
+                        .HasColumnType("bit");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
-                    b.Property<string>("Name");
+                    b.Property<string>("Name")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -727,16 +905,21 @@ namespace Liuliu.Demo.Web.Migrations
             modelBuilder.Entity("OSharp.Core.Systems.KeyValue", b =>
                 {
                     b.Property<Guid>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("uniqueidentifier");
 
-                    b.Property<bool>("IsLocked");
+                    b.Property<bool>("IsLocked")
+                        .HasColumnType("bit");
 
                     b.Property<string>("Key")
-                        .IsRequired();
+                        .IsRequired()
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("ValueJson");
+                    b.Property<string>("ValueJson")
+                        .HasColumnType("nvarchar(max)");
 
-                    b.Property<string>("ValueType");
+                    b.Property<string>("ValueType")
+                        .HasColumnType("nvarchar(max)");
 
                     b.HasKey("Id");
 
@@ -766,19 +949,20 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany()
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.Organization", b =>
                 {
-                    b.HasOne("Liuliu.Demo.Identity.Entities.Organization")
+                    b.HasOne("Liuliu.Demo.Identity.Entities.Organization", null)
                         .WithMany()
                         .HasForeignKey("ParentId");
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.Role", b =>
                 {
-                    b.HasOne("Liuliu.Demo.Infos.Entities.Message")
+                    b.HasOne("Liuliu.Demo.Infos.Entities.Message", null)
                         .WithMany("PublicRoles")
                         .HasForeignKey("MessageId");
                 });
@@ -788,12 +972,13 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.Role", "Role")
                         .WithMany("RoleClaims")
                         .HasForeignKey("RoleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.User", b =>
                 {
-                    b.HasOne("Liuliu.Demo.Infos.Entities.Message")
+                    b.HasOne("Liuliu.Demo.Infos.Entities.Message", null)
                         .WithMany("Recipients")
                         .HasForeignKey("MessageId");
                 });
@@ -803,7 +988,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany("UserClaims")
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserDetail", b =>
@@ -811,7 +997,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithOne("UserDetail")
                         .HasForeignKey("Liuliu.Demo.Identity.Entities.UserDetail", "UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserLogin", b =>
@@ -819,7 +1006,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany("UserLogins")
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserRole", b =>
@@ -827,12 +1015,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.Role", "Role")
                         .WithMany("UserRoles")
                         .HasForeignKey("RoleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany("UserRoles")
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Identity.Entities.UserToken", b =>
@@ -840,7 +1030,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany("UserTokens")
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Infos.Entities.Message", b =>
@@ -848,7 +1039,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "Sender")
                         .WithMany()
                         .HasForeignKey("SenderId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Infos.Entities.MessageReceive", b =>
@@ -856,12 +1048,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Infos.Entities.Message", "Message")
                         .WithMany("Receives")
                         .HasForeignKey("MessageId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany()
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Restrict);
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Infos.Entities.MessageReply", b =>
@@ -869,22 +1063,26 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Infos.Entities.Message", "BelongMessage")
                         .WithMany()
                         .HasForeignKey("BelongMessageId")
-                        .OnDelete(DeleteBehavior.Restrict);
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Infos.Entities.Message", "ParentMessage")
                         .WithMany("Replies")
                         .HasForeignKey("ParentMessageId")
-                        .OnDelete(DeleteBehavior.Restrict);
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Infos.Entities.MessageReply", "ParentReply")
                         .WithMany("Replies")
                         .HasForeignKey("ParentReplyId")
-                        .OnDelete(DeleteBehavior.Restrict);
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany()
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Restrict);
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.EntityRole", b =>
@@ -892,12 +1090,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("OSharp.Core.EntityInfos.EntityInfo", "EntityInfo")
                         .WithMany()
                         .HasForeignKey("EntityId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.Role", "Role")
                         .WithMany()
                         .HasForeignKey("RoleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.EntityUser", b =>
@@ -905,12 +1105,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("OSharp.Core.EntityInfos.EntityInfo", "EntityInfo")
                         .WithMany()
                         .HasForeignKey("EntityId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany()
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.Module", b =>
@@ -925,12 +1127,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("OSharp.Core.Functions.Function", "Function")
                         .WithMany()
                         .HasForeignKey("FunctionId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Security.Entities.Module", "Module")
                         .WithMany()
                         .HasForeignKey("ModuleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.ModuleRole", b =>
@@ -938,12 +1142,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Security.Entities.Module", "Module")
                         .WithMany()
                         .HasForeignKey("ModuleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.Role", "Role")
                         .WithMany()
                         .HasForeignKey("RoleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Security.Entities.ModuleUser", b =>
@@ -951,12 +1157,14 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Security.Entities.Module", "Module")
                         .WithMany()
                         .HasForeignKey("ModuleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.HasOne("Liuliu.Demo.Identity.Entities.User", "User")
                         .WithMany()
                         .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Systems.Entities.AuditEntity", b =>
@@ -964,7 +1172,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Systems.Entities.AuditOperation", "Operation")
                         .WithMany("AuditEntities")
                         .HasForeignKey("OperationId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 
             modelBuilder.Entity("Liuliu.Demo.Systems.Entities.AuditProperty", b =>
@@ -972,7 +1181,8 @@ namespace Liuliu.Demo.Web.Migrations
                     b.HasOne("Liuliu.Demo.Systems.Entities.AuditEntity", "AuditEntity")
                         .WithMany("Properties")
                         .HasForeignKey("AuditEntityId")
-                        .OnDelete(DeleteBehavior.Cascade);
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
                 });
 #pragma warning restore 612, 618
         }
