@@ -1,5 +1,4 @@
 ﻿using System;
-using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Migrations;
 
 namespace Liuliu.Demo.Web.Migrations
@@ -90,7 +89,7 @@ namespace Liuliu.Demo.Web.Migrations
                 columns: table => new
                 {
                     Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                        .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(nullable: false),
                     Remark = table.Column<string>(nullable: true),
                     Code = table.Column<string>(nullable: false),
@@ -114,7 +113,7 @@ namespace Liuliu.Demo.Web.Migrations
                 columns: table => new
                 {
                     Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                        .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(nullable: false),
                     Remark = table.Column<string>(nullable: true),
                     ParentId = table.Column<int>(nullable: true)
@@ -128,61 +127,6 @@ namespace Liuliu.Demo.Web.Migrations
                         principalTable: "Organization",
                         principalColumn: "Id",
                         onDelete: ReferentialAction.Restrict);
-                });
-
-            migrationBuilder.CreateTable(
-                name: "Role",
-                columns: table => new
-                {
-                    Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
-                    Name = table.Column<string>(nullable: false),
-                    NormalizedName = table.Column<string>(nullable: false),
-                    ConcurrencyStamp = table.Column<string>(nullable: true),
-                    Remark = table.Column<string>(maxLength: 512, nullable: true),
-                    IsAdmin = table.Column<bool>(nullable: false),
-                    IsDefault = table.Column<bool>(nullable: false),
-                    IsSystem = table.Column<bool>(nullable: false),
-                    IsLocked = table.Column<bool>(nullable: false),
-                    CreatedTime = table.Column<DateTime>(nullable: false),
-                    DeletedTime = table.Column<DateTime>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_Role", x => x.Id);
-                });
-
-            migrationBuilder.CreateTable(
-                name: "User",
-                columns: table => new
-                {
-                    Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
-                    UserName = table.Column<string>(nullable: false),
-                    NormalizedUserName = table.Column<string>(nullable: false),
-                    NickName = table.Column<string>(nullable: true),
-                    Email = table.Column<string>(nullable: true),
-                    NormalizeEmail = table.Column<string>(nullable: true),
-                    EmailConfirmed = table.Column<bool>(nullable: false),
-                    PasswordHash = table.Column<string>(nullable: true),
-                    HeadImg = table.Column<string>(nullable: true),
-                    SecurityStamp = table.Column<string>(nullable: true),
-                    ConcurrencyStamp = table.Column<string>(nullable: true),
-                    PhoneNumber = table.Column<string>(nullable: true),
-                    PhoneNumberConfirmed = table.Column<bool>(nullable: false),
-                    TwoFactorEnabled = table.Column<bool>(nullable: false),
-                    LockoutEnd = table.Column<DateTimeOffset>(nullable: true),
-                    LockoutEnabled = table.Column<bool>(nullable: false),
-                    AccessFailedCount = table.Column<int>(nullable: false),
-                    IsSystem = table.Column<bool>(nullable: false),
-                    IsLocked = table.Column<bool>(nullable: false),
-                    CreatedTime = table.Column<DateTime>(nullable: false),
-                    DeletedTime = table.Column<DateTime>(nullable: true),
-                    Remark = table.Column<string>(nullable: true)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_User", x => x.Id);
                 });
 
             migrationBuilder.CreateTable(
@@ -233,6 +177,29 @@ namespace Liuliu.Demo.Web.Migrations
                 });
 
             migrationBuilder.CreateTable(
+                name: "AuditProperty",
+                columns: table => new
+                {
+                    Id = table.Column<Guid>(nullable: false),
+                    DisplayName = table.Column<string>(nullable: true),
+                    FieldName = table.Column<string>(nullable: true),
+                    OriginalValue = table.Column<string>(nullable: true),
+                    NewValue = table.Column<string>(nullable: true),
+                    DataType = table.Column<string>(nullable: true),
+                    AuditEntityId = table.Column<Guid>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_AuditProperty", x => x.Id);
+                    table.ForeignKey(
+                        name: "FK_AuditProperty_AuditEntity_AuditEntityId",
+                        column: x => x.AuditEntityId,
+                        principalTable: "AuditEntity",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Cascade);
+                });
+
+            migrationBuilder.CreateTable(
                 name: "EntityRole",
                 columns: table => new
                 {
@@ -253,12 +220,93 @@ namespace Liuliu.Demo.Web.Migrations
                         principalTable: "EntityInfo",
                         principalColumn: "Id",
                         onDelete: ReferentialAction.Cascade);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "EntityUser",
+                columns: table => new
+                {
+                    Id = table.Column<Guid>(nullable: false),
+                    UserId = table.Column<int>(nullable: false),
+                    EntityId = table.Column<Guid>(nullable: false),
+                    FilterGroupJson = table.Column<string>(nullable: true),
+                    IsLocked = table.Column<bool>(nullable: false),
+                    CreatedTime = table.Column<DateTime>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_EntityUser", x => x.Id);
                     table.ForeignKey(
-                        name: "FK_EntityRole_Role_RoleId",
-                        column: x => x.RoleId,
-                        principalTable: "Role",
+                        name: "FK_EntityUser_EntityInfo_EntityId",
+                        column: x => x.EntityId,
+                        principalTable: "EntityInfo",
                         principalColumn: "Id",
                         onDelete: ReferentialAction.Cascade);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "MessageReceive",
+                columns: table => new
+                {
+                    Id = table.Column<Guid>(nullable: false),
+                    ReadDate = table.Column<DateTime>(nullable: false),
+                    NewReplyCount = table.Column<int>(nullable: false),
+                    CreatedTime = table.Column<DateTime>(nullable: false),
+                    MessageId = table.Column<Guid>(nullable: false),
+                    UserId = table.Column<int>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_MessageReceive", x => x.Id);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "MessageReply",
+                columns: table => new
+                {
+                    Id = table.Column<Guid>(nullable: false),
+                    Content = table.Column<string>(nullable: false),
+                    IsRead = table.Column<bool>(nullable: false),
+                    ParentMessageId = table.Column<Guid>(nullable: false),
+                    ParentReplyId = table.Column<Guid>(nullable: false),
+                    IsLocked = table.Column<bool>(nullable: false),
+                    DeletedTime = table.Column<DateTime>(nullable: true),
+                    CreatedTime = table.Column<DateTime>(nullable: false),
+                    UserId = table.Column<int>(nullable: false),
+                    BelongMessageId = table.Column<Guid>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_MessageReply", x => x.Id);
+                    table.ForeignKey(
+                        name: "FK_MessageReply_MessageReply_ParentReplyId",
+                        column: x => x.ParentReplyId,
+                        principalTable: "MessageReply",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Role",
+                columns: table => new
+                {
+                    Id = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:Identity", "1, 1"),
+                    Name = table.Column<string>(nullable: false),
+                    NormalizedName = table.Column<string>(nullable: false),
+                    ConcurrencyStamp = table.Column<string>(nullable: true),
+                    Remark = table.Column<string>(maxLength: 512, nullable: true),
+                    IsAdmin = table.Column<bool>(nullable: false),
+                    IsDefault = table.Column<bool>(nullable: false),
+                    IsSystem = table.Column<bool>(nullable: false),
+                    IsLocked = table.Column<bool>(nullable: false),
+                    CreatedTime = table.Column<DateTime>(nullable: false),
+                    DeletedTime = table.Column<DateTime>(nullable: true),
+                    MessageId = table.Column<Guid>(nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Role", x => x.Id);
                 });
 
             migrationBuilder.CreateTable(
@@ -291,7 +339,7 @@ namespace Liuliu.Demo.Web.Migrations
                 columns: table => new
                 {
                     Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                        .Annotation("SqlServer:Identity", "1, 1"),
                     RoleId = table.Column<int>(nullable: false),
                     ClaimType = table.Column<string>(nullable: true),
                     ClaimValue = table.Column<string>(nullable: true)
@@ -308,31 +356,37 @@ namespace Liuliu.Demo.Web.Migrations
                 });
 
             migrationBuilder.CreateTable(
-                name: "EntityUser",
+                name: "User",
                 columns: table => new
                 {
-                    Id = table.Column<Guid>(nullable: false),
-                    UserId = table.Column<int>(nullable: false),
-                    EntityId = table.Column<Guid>(nullable: false),
-                    FilterGroupJson = table.Column<string>(nullable: true),
+                    Id = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:Identity", "1, 1"),
+                    UserName = table.Column<string>(nullable: false),
+                    NormalizedUserName = table.Column<string>(nullable: false),
+                    NickName = table.Column<string>(nullable: true),
+                    Email = table.Column<string>(nullable: true),
+                    NormalizeEmail = table.Column<string>(nullable: true),
+                    EmailConfirmed = table.Column<bool>(nullable: false),
+                    PasswordHash = table.Column<string>(nullable: true),
+                    HeadImg = table.Column<string>(nullable: true),
+                    SecurityStamp = table.Column<string>(nullable: true),
+                    ConcurrencyStamp = table.Column<string>(nullable: true),
+                    PhoneNumber = table.Column<string>(nullable: true),
+                    PhoneNumberConfirmed = table.Column<bool>(nullable: false),
+                    TwoFactorEnabled = table.Column<bool>(nullable: false),
+                    LockoutEnd = table.Column<DateTimeOffset>(nullable: true),
+                    LockoutEnabled = table.Column<bool>(nullable: false),
+                    AccessFailedCount = table.Column<int>(nullable: false),
+                    IsSystem = table.Column<bool>(nullable: false),
                     IsLocked = table.Column<bool>(nullable: false),
-                    CreatedTime = table.Column<DateTime>(nullable: false)
+                    CreatedTime = table.Column<DateTime>(nullable: false),
+                    DeletedTime = table.Column<DateTime>(nullable: true),
+                    Remark = table.Column<string>(nullable: true),
+                    MessageId = table.Column<Guid>(nullable: true)
                 },
                 constraints: table =>
                 {
-                    table.PrimaryKey("PK_EntityUser", x => x.Id);
-                    table.ForeignKey(
-                        name: "FK_EntityUser_EntityInfo_EntityId",
-                        column: x => x.EntityId,
-                        principalTable: "EntityInfo",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                    table.ForeignKey(
-                        name: "FK_EntityUser_User_UserId",
-                        column: x => x.UserId,
-                        principalTable: "User",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
+                    table.PrimaryKey("PK_User", x => x.Id);
                 });
 
             migrationBuilder.CreateTable(
@@ -352,6 +406,35 @@ namespace Liuliu.Demo.Web.Migrations
                     table.ForeignKey(
                         name: "FK_LoginLog_User_UserId",
                         column: x => x.UserId,
+                        principalTable: "User",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Cascade);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Message",
+                columns: table => new
+                {
+                    Id = table.Column<Guid>(nullable: false),
+                    Title = table.Column<string>(nullable: false),
+                    Content = table.Column<string>(nullable: false),
+                    MessageType = table.Column<int>(nullable: false),
+                    NewReplyCount = table.Column<int>(nullable: false),
+                    IsSended = table.Column<bool>(nullable: false),
+                    CanReply = table.Column<bool>(nullable: false),
+                    BeginDate = table.Column<DateTime>(nullable: true),
+                    EndDate = table.Column<DateTime>(nullable: true),
+                    IsLocked = table.Column<bool>(nullable: false),
+                    DeletedTime = table.Column<DateTime>(nullable: true),
+                    CreatedTime = table.Column<DateTime>(nullable: false),
+                    SenderId = table.Column<int>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Message", x => x.Id);
+                    table.ForeignKey(
+                        name: "FK_Message_User_SenderId",
+                        column: x => x.SenderId,
                         principalTable: "User",
                         principalColumn: "Id",
                         onDelete: ReferentialAction.Cascade);
@@ -388,7 +471,7 @@ namespace Liuliu.Demo.Web.Migrations
                 columns: table => new
                 {
                     Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                        .Annotation("SqlServer:Identity", "1, 1"),
                     UserId = table.Column<int>(nullable: false),
                     ClaimType = table.Column<string>(nullable: false),
                     ClaimValue = table.Column<string>(nullable: true)
@@ -409,7 +492,7 @@ namespace Liuliu.Demo.Web.Migrations
                 columns: table => new
                 {
                     Id = table.Column<int>(nullable: false)
-                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                        .Annotation("SqlServer:Identity", "1, 1"),
                     RegisterIp = table.Column<string>(nullable: true),
                     UserId = table.Column<int>(nullable: false)
                 },
@@ -496,29 +579,6 @@ namespace Liuliu.Demo.Web.Migrations
                         onDelete: ReferentialAction.Cascade);
                 });
 
-            migrationBuilder.CreateTable(
-                name: "AuditProperty",
-                columns: table => new
-                {
-                    Id = table.Column<Guid>(nullable: false),
-                    DisplayName = table.Column<string>(nullable: true),
-                    FieldName = table.Column<string>(nullable: true),
-                    OriginalValue = table.Column<string>(nullable: true),
-                    NewValue = table.Column<string>(nullable: true),
-                    DataType = table.Column<string>(nullable: true),
-                    AuditEntityId = table.Column<Guid>(nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_AuditProperty", x => x.Id);
-                    table.ForeignKey(
-                        name: "FK_AuditProperty_AuditEntity_AuditEntityId",
-                        column: x => x.AuditEntityId,
-                        principalTable: "AuditEntity",
-                        principalColumn: "Id",
-                        onDelete: ReferentialAction.Cascade);
-                });
-
             migrationBuilder.InsertData(
                 table: "KeyValue",
                 columns: new[] { "Id", "IsLocked", "Key", "ValueJson", "ValueType" },
@@ -535,8 +595,8 @@ namespace Liuliu.Demo.Web.Migrations
 
             migrationBuilder.InsertData(
                 table: "Role",
-                columns: new[] { "Id", "ConcurrencyStamp", "CreatedTime", "DeletedTime", "IsAdmin", "IsDefault", "IsLocked", "IsSystem", "Name", "NormalizedName", "Remark" },
-                values: new object[] { 1, "97313840-7874-47e5-81f2-565613c8cdcc", new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified), null, true, false, false, true, "系统管理员", "系统管理员", "系统最高权限管理角色" });
+                columns: new[] { "Id", "ConcurrencyStamp", "CreatedTime", "DeletedTime", "IsAdmin", "IsDefault", "IsLocked", "IsSystem", "MessageId", "Name", "NormalizedName", "Remark" },
+                values: new object[] { 1, "97313840-7874-47e5-81f2-565613c8cdcc", new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified), null, true, false, false, true, null, "系统管理员", "系统管理员", "系统最高权限管理角色" });
 
             migrationBuilder.CreateIndex(
                 name: "IX_AuditEntity_OperationId",
@@ -588,6 +648,41 @@ namespace Liuliu.Demo.Web.Migrations
                 column: "UserId");
 
             migrationBuilder.CreateIndex(
+                name: "IX_Message_SenderId",
+                table: "Message",
+                column: "SenderId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_MessageReceive_MessageId",
+                table: "MessageReceive",
+                column: "MessageId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_MessageReceive_UserId",
+                table: "MessageReceive",
+                column: "UserId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_MessageReply_BelongMessageId",
+                table: "MessageReply",
+                column: "BelongMessageId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_MessageReply_ParentMessageId",
+                table: "MessageReply",
+                column: "ParentMessageId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_MessageReply_ParentReplyId",
+                table: "MessageReply",
+                column: "ParentReplyId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_MessageReply_UserId",
+                table: "MessageReply",
+                column: "UserId");
+
+            migrationBuilder.CreateIndex(
                 name: "IX_Module_ParentId",
                 table: "Module",
                 column: "ParentId");
@@ -631,6 +726,11 @@ namespace Liuliu.Demo.Web.Migrations
                 column: "ParentId");
 
             migrationBuilder.CreateIndex(
+                name: "IX_Role_MessageId",
+                table: "Role",
+                column: "MessageId");
+
+            migrationBuilder.CreateIndex(
                 name: "RoleNameIndex",
                 table: "Role",
                 columns: new[] { "NormalizedName", "DeletedTime" },
@@ -641,6 +741,11 @@ namespace Liuliu.Demo.Web.Migrations
                 name: "IX_RoleClaim_RoleId",
                 table: "RoleClaim",
                 column: "RoleId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_User_MessageId",
+                table: "User",
+                column: "MessageId");
 
             migrationBuilder.CreateIndex(
                 name: "EmailIndex",
@@ -695,10 +800,86 @@ namespace Liuliu.Demo.Web.Migrations
                 columns: new[] { "UserId", "LoginProvider", "Name" },
                 unique: true,
                 filter: "[LoginProvider] IS NOT NULL AND [Name] IS NOT NULL");
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_EntityRole_Role_RoleId",
+                table: "EntityRole",
+                column: "RoleId",
+                principalTable: "Role",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Cascade);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_EntityUser_User_UserId",
+                table: "EntityUser",
+                column: "UserId",
+                principalTable: "User",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Cascade);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_MessageReceive_User_UserId",
+                table: "MessageReceive",
+                column: "UserId",
+                principalTable: "User",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Restrict);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_MessageReceive_Message_MessageId",
+                table: "MessageReceive",
+                column: "MessageId",
+                principalTable: "Message",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Cascade);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_MessageReply_User_UserId",
+                table: "MessageReply",
+                column: "UserId",
+                principalTable: "User",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Restrict);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_MessageReply_Message_BelongMessageId",
+                table: "MessageReply",
+                column: "BelongMessageId",
+                principalTable: "Message",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Restrict);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_MessageReply_Message_ParentMessageId",
+                table: "MessageReply",
+                column: "ParentMessageId",
+                principalTable: "Message",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Restrict);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_Role_Message_MessageId",
+                table: "Role",
+                column: "MessageId",
+                principalTable: "Message",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Restrict);
+
+            migrationBuilder.AddForeignKey(
+                name: "FK_User_Message_MessageId",
+                table: "User",
+                column: "MessageId",
+                principalTable: "Message",
+                principalColumn: "Id",
+                onDelete: ReferentialAction.Restrict);
         }
 
         protected override void Down(MigrationBuilder migrationBuilder)
         {
+            migrationBuilder.DropForeignKey(
+                name: "FK_Message_User_SenderId",
+                table: "Message");
+
             migrationBuilder.DropTable(
                 name: "AuditProperty");
 
@@ -713,6 +894,12 @@ namespace Liuliu.Demo.Web.Migrations
 
             migrationBuilder.DropTable(
                 name: "LoginLog");
+
+            migrationBuilder.DropTable(
+                name: "MessageReceive");
+
+            migrationBuilder.DropTable(
+                name: "MessageReply");
 
             migrationBuilder.DropTable(
                 name: "ModuleFunction");
@@ -760,10 +947,13 @@ namespace Liuliu.Demo.Web.Migrations
                 name: "Role");
 
             migrationBuilder.DropTable(
+                name: "AuditOperation");
+
+            migrationBuilder.DropTable(
                 name: "User");
 
             migrationBuilder.DropTable(
-                name: "AuditOperation");
+                name: "Message");
         }
     }
 }
