@@ -12,6 +12,8 @@ using OSharp.AspNetCore.SignalR;
 using System;
 using System.ComponentModel;
 
+using Microsoft.AspNetCore.Routing;
+
 
 namespace Liuliu.Demo.Web.Startups
 {
@@ -22,16 +24,16 @@ namespace Liuliu.Demo.Web.Startups
     public class SignalRPack : SignalRPackBase
     {
         /// <summary>
-        /// 重写以获取Hub路由创建委托
+        /// 重写以获取Hub终结点路由创建委托
         /// </summary>
         /// <param name="serviceProvider">服务提供者</param>
         /// <returns></returns>
-        protected override Action<HubRouteBuilder> GetHubRouteBuildAction(IServiceProvider serviceProvider)
+        protected override Action<IEndpointRouteBuilder> GetEndpointRouteBuilderAction(IServiceProvider serviceProvider)
         {
-            return new Action<HubRouteBuilder>(builder =>
+            return new Action<IEndpointRouteBuilder>(endpoints =>
             {
                 // 在这实现Hub的路由映射
-                // 例如：builder.MapHub<MyHub>();
+                // 例如：endpoints.MapHub<ChatHub>();
             });
         }
     }
